% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-autoQC01.R
\name{autoQC01Server}
\alias{autoQC01Server}
\title{autoQC01Server shiny module}
\usage{
autoQC01Server(id, filterValues, BFabric, inputfile)
}
\arguments{
\item{id}{}

\item{filterValues}{time and instrument information}
}
\value{
data.frame
}
\description{
autoQC01Server shiny module
}
