% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-autoQC03.R
\name{autoQC03Server}
\alias{autoQC03Server}
\title{autoQC01Server shiny module}
\usage{
autoQC03Server(
  id,
  filterValues,
  BFabric,
  inputfile,
  readFUN,
  ggplot2FUN,
  title,
  footer
)
}
\arguments{
\item{id}{}

\item{filterValues}{time and instrument information}
}
\description{
handles DDA and DIA
}
