% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.assignInstrument}
\alias{.assignInstrument}
\title{Assign instruments of a given FGCZ file path}
\usage{
.assignInstrument(x, coln = "File.Name")
}
\arguments{
\item{x}{a data frame with a column 'File.Name'}

\item{coln}{the column name of the file name}
}
\value{
a data.frame with assigned instruments
}
\description{
Assign instruments of a given FGCZ file path
}
