% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.determineLastEntry}
\alias{.determineLastEntry}
\title{determine last entry of a data.frame}
\usage{
.determineLastEntry(x)
}
\description{
data are spitted bu Instrument and the lastest entry of each instrument is returned
}
\examples{
.composeInstrumentEvents() |> .determineLastEntry()
}
