% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-rawrr.R
\name{.isPeak}
\alias{.isPeak}
\title{When is a peak a peak?}
\usage{
.isPeak(FWHM, x)
}
\arguments{
\item{FWHM}{proposal}

\item{x}{curve}
}
\value{
TRUE if at least one point is in between the FWHM x-range.
}
\description{
A super basic heuristic to descide if x describes a Gaussian peak.
}
