% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-readr.R
\name{.readAutoQC01}
\alias{.readAutoQC01}
\title{.readAutoQC01('/Users/cp/Downloads/dump/autoQC01-fit-apex-auc-fwhm.txt') -> S
lattice::xyplot(value ~ time | Instrument * variable, group = peptide, data = S, scales = list(y = list(relation = "free")), pch = '.')}
\usage{
.readAutoQC01(filename)
}
\description{
.readAutoQC01('/Users/cp/Downloads/dump/autoQC01-fit-apex-auc-fwhm.txt') -> S
lattice::xyplot(value ~ time | Instrument * variable, group = peptide, data = S, scales = list(y = list(relation = "free")), pch = '.')
}
