% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-readr.R
\name{.readComet}
\alias{.readComet}
\title{read comet qc data from RData file}
\usage{
.readComet(filename)
}
\arguments{
\item{filename}{}
}
\value{
a data.frame object
}
\description{
read comet qc data from RData file
}
\examples{
.readComet('~/Downloads/dump/comet.RData') -> S
}
