% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIMSdda-comet-summary.R
\name{harmonizeTimsDDA}
\alias{harmonizeTimsDDA}
\title{Harmonize comet DDA searches from Bruker TIMS devices}
\usage{
harmonizeTimsDDA(input)
}
\value{
a data.frame with the following to be used for the qc dashboard
}
\description{
Harmonize comet DDA searches from Bruker TIMS devices
}
