% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-rawrr.R
\name{rawrrServer}
\alias{rawrrServer}
\title{Extracts and plots chromatographic profile of a given m/z vector using the 
rawrr Rpkg from Bioconductor.}
\usage{
rawrrServer(id, vals)
}
\arguments{
\item{id}{shiny namespace id}

\item{vals}{a reactive variable containing fn, the rawfile, and a numeric
vector of mZ values. The names of the vector, e.g., AA sequeces, are used as
labels.#'}
}
\value{
nothing
}
\description{
Extracts and plots chromatographic profile of a given m/z vector using the 
rawrr Rpkg from Bioconductor.
}
